[iOS] feat: MenuGenrator 생성
